\name{sticklebacks}
\Rdversion{1.1}
\alias{sticklebacks}
\docType{data}
\title{
Data from Nordic ninespine sticklebacks (Pungitius pungitius).
}
\description{
183 fish from 4 local populations sampled for 12 microsatellite loci.
}
\usage{data(sticklebacks)}
\format{
  A data frame with 183 observations on the following 25 variables.
  \describe{
    \item{\code{subpop}}{subpopulation code}
    \item{\code{X1125P.1}}{allelic copy 1, marker X1125P}
    \item{\code{X1125P.2}}{allelic copy 2, marker X1125P}
    \item{\code{X4174P.1}}{allelic copy 1, marker X4174P}
    \item{\code{X4174P.2}}{allelic copy 2, marker X4174P}
    \item{\code{X7033P.1}}{allelic copy 1, marker X7033P}
    \item{\code{X7033P.2}}{. . .}
    \item{\code{Stn49.1}}{. . .}
    \item{\code{Stn49.2}}{. . .}
    \item{\code{Stn96.1}}{. . .}
    \item{\code{Stn96.2}}{. . .}
    \item{\code{Stn100.1}}{. . .}
    \item{\code{Stn100.2}}{. . .}
    \item{\code{Stn130.1}}{. . .}
    \item{\code{Stn130.2}}{. . .}
    \item{\code{Stn163.1}}{. . .}
    \item{\code{Stn163.2}}{. . .}
    \item{\code{Stn173.1}}{. . .}
    \item{\code{Stn173.2}}{. . .}
    \item{\code{Stn196.1}}{. . .}
    \item{\code{Stn196.2}}{. . .}
    \item{\code{Stn198.1}}{. . .}
    \item{\code{STn198.2}}{. . .}
    \item{\code{Stn380.1}}{. . .}
    \item{\code{Stn380.2}}{. . .}    
  }
}
\details{
	Each row represents one individual.
	In this case, the genotypes are microsatellite fragment lengths.
	Missing data is coded as NA.
}
\source{
Shikano T, Shimada Y, Herczeg G and Merila J (2010): History vs. habitat type: explaining the genetic structure of European nine-spined stickleback (Pungitius pungitius) populations. Molecular Ecology 19: 1147-1161.
}

\examples{
data(sticklebacks)
}
\keyword{datasets}
